backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
redis[hiredis]==5.0.0
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.30
psycopg2-binary==2.9.9
pydantic==2.7.0
prometheus-client==0.20.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.27.0
